extract solanaEncode
